Fix "size" URL parameter conversion to be int value or "full" literal.
CacheManager expects value of `size` parameter in it's methods to be either
`int` or `str` with value _"full"_ (this was infered from
`config['ALLOWED_SIZES']`).
The problem was that values that seems to be valid (like `str` _"200"_) weren't
converted to integers and later failed the check
`size in config['ALLOWED_SIZES']`. Fixed by using standard flask parameter
type conversion with extra route config for value _"full"_.